backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
supabase==2.10.0
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.2
